vuln-analyzer-agent/pyproject.toml
[project]
name = "vuln-analyzer"
version = "1.0.0"
description = "Python agent for analyzing vulnerabilities in Java Spring Maven projects"
authors = [{name = "Developer", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "lxml>=5.1.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
mend = ["ws-sdk>=24.3.1"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
vuln-analyzer = "vuln_analyzer.main:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
